Divide edit method into two possible renderings
1. The editable field we have been rendering on edit
2. The front-end markup we render in the save method

Our conditional checks if the block is currently, selected.
If it is selected, we display the editable field we have been using.

However, if the block is NOT selected, we render it the same way we do
when saving (i.e. the same way it is displayed on the front-end).